+++ mod/common/src/main/java/io/github/fishstiz/packed_packs/config/Config.java
@@ -76,7 +76,13 @@ public final class Config {
     }
 
     public void setSort(SortOption sort) {
-        this.sort = sort.name();
+        if (sort instanceof SortOption.Locked(SortOption locked)) {
+            if (locked != null) {
+                this.sort = locked.name();
+            }
+        } else {
+            this.sort = sort.name();
+        }
     }
 
     public ResourcePacks getResourcepacks() {