+++ mod/common/src/main/java/io/github/fishstiz/packed_packs/config/JsonLoader.java
@@ -57,15 +57,21 @@ public final class JsonLoader {
         }
     }
 
-    public static <T> boolean saveJson(T serializable, Path path) {
+    public static boolean saveJson(Object serializable, Path path, boolean createDirs) {
         try {
             String json = GSON.toJson(serializable);
-            Files.createDirectories(path.getParent());
+            if (createDirs) {
+                Files.createDirectories(path.getParent());
+            }
             Files.writeString(path, json, StandardCharsets.UTF_8, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
             return true;
         } catch (Exception e) {
             PackedPacks.LOGGER.info("[packed_packs] Failed to save file at '{}'.", path, e);
             return false;
         }
+    }
+
+    public static <T> boolean saveJson(T serializable, Path path) {
+        return saveJson(serializable, path, true);
     }
 }